backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
aiofiles
pytest
httpx
pytest-asyncio
